feature-gitcommitter auto-copy commits to new path
.\gitCommitter.ps1 is the command I actually ran. It has better handling,
the beginnings of a function, and descriptive output.

.\gc.ps1 is a minimal version which shows the plain logic a bit more
clearly. I didn't run this one because the longer one is more efficient
and this one is a little hacked. It displays no output and is more to help
see the logic plainly when reading source. While testing, I got the
logic down to as little as 28 lines if I gave up constant variables, etc.

.\outputLog.txt is a log of the status before, the output of
.\gitCommitter.ps1, the log after, and the status after.